Moved transform.py to sim/
