--- motor.py
@@ -1,2 +1,0 @@
-import numpy as np
-
